+++ HISTORY.rst
@@ -2,6 +2,16 @@
 History
 =======
 
+0.3.4 (Future)
+^^^^^^^^^^^^^^^^^^
+
+* Fix primitive type with metadata handling.
+* Fix multi-dimensional array handling.
+* Allow empty constructor in generated objects.
+* Support encoding bytes in JSON as base64 string.
+* Added encoding option for source and target files.
+* Support arrays with union item types.
+
 0.3.3 (2022-01-26)
 ^^^^^^^^^^^^^^^^^^